@@ -33,33 +33,24 @@
     "@auth/dynamodb-adapter": "^1.0.3",
     "@aws-sdk/lib-dynamodb": "^3.428.0",
     "@chakra-ui/icons": "^2.0.16",
+    "@chakra-ui/next-js": "^2.1.5",
     "@chakra-ui/react": "^2.4.6",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@reduxjs/toolkit": "^1.9.1",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
     "chakra-react-select": "^4.6.0",
     "date-fns": "^2.30.0",
     "dynamoose": "^3.2.0",
-    "eslint": "^8.31.0",
-    "eslint-config-next": "^13.2.4",
     "final-form": "^4.20.9",
     "final-form-calculate": "^1.3.2",
     "framer-motion": "^6",
-    "google-auth-library": "^8.7.0",
-    "googleapis": "^110.0.0",
-    "i18next": "^22.4.9",
-    "jwt-decode": "^3.1.2",
     "loadash": "^1.0.0",
-    "next": "^13.2.4",
-    "next-auth": "^4.23.2",
-    "next-i18next": "^13.0.3",
+    "next": "^14.0.0",
+    "next-auth": "^4.24.3",
+    "next-international": "^1.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-final-form": "^6.5.9",
-    "react-i18next": "^12.1.4",
     "react-icons": "^4.7.1",
     "react-imask": "^6.6.0",
     "react-redux": "^8.0.5",
@@ -67,23 +58,28 @@
     "redux-persist": "^6.0.0",
     "redux-state-sync": "^3.1.4",
     "slugify": "^1.6.6",
-    "typescript": "4.9.4",
+    "typescript": "^5.2.2",
     "ulid": "^2.3.0",
-    "uuid": "^9.0.1",
     "yup": "^1.2.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.4.4",
     "@commitlint/config-conventional": "^17.4.4",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
     "@types/google.accounts": "^0.0.5",
+    "@types/jest": "^29.5.6",
+    "@types/node": "^20.8.9",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
     "@types/react-speech-recognition": "^3.9.2",
     "@types/react-table": "^7.7.14",
     "@types/redux-state-sync": "^3.1.5",
     "@types/uuid": "^9.0.3",
     "@typescript-eslint/eslint-plugin": "^5.48.0",
     "@typescript-eslint/parser": "^5.48.0",
+    "eslint": "^8.52.0",
+    "eslint-config-next": "^14.0.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-import-helpers": "^1.3.1",
     "eslint-plugin-jest": "^27.2.0",
@@ -91,8 +87,8 @@
     "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
-    "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "prettier": "^2.8.1",
     "ts-node": "^10.9.1"
   }
