@@ -1,16 +1,16 @@
+'use client';
 import { useMemo } from 'react';
 
-import { useRouter } from 'next/router';
-import { useTranslation } from 'react-i18next';
+import { usePathname } from 'next/navigation';
+
+import { useScopedI18n } from '~/i18n/client';
 
 import { MENU_ITEMS } from './const';
 import { HeaderItemProps, UseHeaderProps } from './types';
 
 export const useHeader: UseHeaderProps = () => {
-  const { pathname } = useRouter();
-  const { t } = useTranslation(undefined, {
-    keyPrefix: 'header.label',
-  });
+  const pathname = usePathname();
+  const t = useScopedI18n('translations.header.label');
 
   const headersItems = useMemo<HeaderItemProps[]>(
     () =>
