@@ -1,14 +1,11 @@
 import { FC, MouseEventHandler, useCallback } from 'react';
 
 import Link from 'next/link';
-import { useTranslation } from 'react-i18next';
 
 import { BreadcrumbStyledItem, BreadcrumbWrapper } from './styles';
 import { BreadcrumbProps } from './types';
 
 export const Breadcrumb: FC<BreadcrumbProps> = ({ items }) => {
-  const { t } = useTranslation();
-
   const stopOnClick = useCallback<MouseEventHandler<HTMLAnchorElement>>(
     (event) => {
       event.preventDefault();
@@ -30,7 +27,7 @@ export const Breadcrumb: FC<BreadcrumbProps> = ({ items }) => {
             href={href}
             onClick={items.length - 1 === index ? stopOnClick : undefined}
           >
-            <span>{text || t(textKey || '')}</span>
+            <span>{text}</span>
           </Link>
         </BreadcrumbStyledItem>
       ))}
