@@ -1,8 +1,10 @@
+import { FlattenLocaleKeys } from '~/i18n/types';
+
 export type BreadcrumbItemProps = {
   href: string;
 } & (
   | { text: string; textKey?: undefined }
-  | { text?: undefined; textKey: string }
+  | { text?: undefined; textKey: FlattenLocaleKeys }
 );
 
 export type BreadcrumbProps = {
