@@ -1,15 +1,16 @@
 import { forwardRef, Button } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
+
+import { useScopedI18n } from '~/i18n/client';
 
 import { VariantButtonProps } from '../types';
 
 export const CancelButton = forwardRef<VariantButtonProps, 'button'>(
   ({ text, children, ...props }, ref) => {
-    const { t } = useTranslation();
+    const t = useScopedI18n('components.buttons');
 
     return (
       <Button variant='outline' {...props} ref={ref}>
-        {text || children || t('words.cancel')}
+        {text || children || t('cancel')}
       </Button>
     );
   }
