@@ -1,5 +1,4 @@
 import { Thead } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
 
 import { TDataId } from '../types';
 import { StyledTheadTh, StyledTheadTr } from './styles';
@@ -8,8 +7,6 @@ import { DataTableHeaderProps } from './types';
 export function DataTableHeader<
   TData extends TDataId = TDataId & Record<string, string>
 >({ columns }: DataTableHeaderProps<TData>) {
-  const { t } = useTranslation();
-
   return (
     <Thead>
       <StyledTheadTr>
@@ -19,9 +16,9 @@ export function DataTableHeader<
               Object.entries(x).filter(([key]) => key !== 'render')
             )
           )
-          .map(({ label, tLabel, ...columnProps }) => (
+          .map(({ label, ...columnProps }) => (
             <StyledTheadTh {...columnProps} key={label}>
-              {label ? label : t(tLabel)}
+              {label}
             </StyledTheadTh>
           ))}
       </StyledTheadTr>
