@@ -1,22 +1,22 @@
 import { FC } from 'react';
 
 import { useFormState } from 'react-final-form';
-import { useTranslation } from 'react-i18next';
+
+import { useScopedI18n } from '~/i18n/client';
 
 import { SaveButton } from '../../Buttons';
 import { TooltipComponent } from '../../TooltipComponent';
 import { FromSaveButtonProps } from './types';
 
 export const FormSaveButton: FC<FromSaveButtonProps> = (props) => {
-  const { t } = useTranslation(undefined, {
-    keyPrefix: 'form',
-  });
+  const t = useScopedI18n('components.form');
+
   const { pristine, hasValidationErrors } = useFormState();
 
-  const options: [boolean, string][] = [
+  const options = [
     [pristine, 'disabledByPristine'],
     [hasValidationErrors, 'disabledByHasValidationErrors'],
-  ];
+  ] as const;
   const [, translateKey] = options.find(([condition]) =>
     Boolean(condition)
   ) || [false, undefined];
