@@ -7,8 +7,8 @@ const AUTHENTICATION_ERROR = {
   authError: 'Cannot be authenticated, authorization header is not filled',
 };
 
-export const getCustomerId: HandleGetCustomerId = async (req, res) => {
-  const session = await getServerSession(req, res, authOptions);
+export const getCustomerId: HandleGetCustomerId = async () => {
+  const session = await getServerSession(authOptions);
   const id = (session?.user as { googleId?: string })?.googleId;
 
   if (!id) return AUTHENTICATION_ERROR;
