@@ -1,11 +1,4 @@
-import { NextApiResponse } from 'next';
-
-import { ReqProps } from '~/core/api';
-
-export type HandleGetCustomerId = (
-  req: ReqProps['req'],
-  res: NextApiResponse
-) => Promise<
+export type HandleGetCustomerId = () => Promise<
   | {
       customerId: string;
       authError: '';
