@@ -1,20 +1,21 @@
-import { EnumHttpStatus, sendError } from '~/core/api';
+import { NextResponse } from 'next/server';
+
+import { sendError } from '~/core/api';
 import { getCustomerId } from '~/core/modules/Customer/auth';
 
 import { Patient } from '../../../api/model';
 import { CaseReport, getCaseReportPK } from '../model';
 import { GetCaseReportHandler } from './types';
 
-export const get: GetCaseReportHandler = async (req, res) => {
-  const { authError, customerId } = await getCustomerId(req, res);
-  if (!customerId) return sendError({ res, error: authError });
+export const get: GetCaseReportHandler = async (req, ctx) => {
+  const { authError, customerId } = await getCustomerId();
+  if (!customerId) return sendError({ error: authError });
 
-  const { patientId, caseReportId } = req.query;
-  if (!patientId)
-    return sendError({ res, error: 'patientId not found in query' });
+  const { patientId, caseReportId } = ctx.params;
+  if (!patientId) return sendError({ error: 'patientId not found in query' });
 
   if (!caseReportId)
-    return sendError({ res, error: 'caseReportId not found in query' });
+    return sendError({ error: 'caseReportId not found in query' });
 
   const caseReportQuery = CaseReport.query({
     PK: getCaseReportPK({ tenantId: customerId, patientId }),
@@ -35,19 +36,11 @@ export const get: GetCaseReportHandler = async (req, res) => {
     patientNameQuery,
   ]);
   if (!caseReport)
-    return sendError({
-      res,
-      error: 'caseReport not found',
-      status: 'NotFound',
-    });
+    return sendError({ error: 'caseReport not found', status: 'NotFound' });
   if (!patientName)
-    return sendError({
-      res,
-      error: 'patientName not found',
-      status: 'NotFound',
-    });
+    return sendError({ error: 'patientName not found', status: 'NotFound' });
 
-  return res.status(EnumHttpStatus.Success).send({
+  return NextResponse.json({
     caseReport: {
       id: caseReport.SK,
       content: caseReport.content,
