@@ -2,10 +2,10 @@ import { FC } from 'react';
 
 import { EditIcon } from '@chakra-ui/icons';
 import { Button } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
 
 import { DataTable, DataTableColumnProps } from '~/common/components/DataTable';
 import { useLoader } from '~/core/services';
+import { useScopedI18n } from '~/i18n/client';
 
 import { ListCaseReportResume } from '../../../../api/list/types';
 import { PatientCaseReportUpsertModal } from '../../../../modal/upsert';
@@ -17,12 +17,7 @@ export const PatientCaseReportTable: FC<PatientCaseReportTableProps> = ({
   patientName,
   caseReports,
 }) => {
-  const { t: tWords } = useTranslation(undefined, {
-    keyPrefix: 'words',
-  });
-  const { t } = useTranslation(undefined, {
-    keyPrefix: 'pages.patient.caseReport',
-  });
+  const t = useScopedI18n('translations.pages.patient.caseReport');
   const { isLoading } = useLoader('DEFAULT');
 
   const COLUMNS: DataTableColumnProps<ListCaseReportResume>[] = [
@@ -30,7 +25,7 @@ export const PatientCaseReportTable: FC<PatientCaseReportTableProps> = ({
       w: '310px',
       maxW: '310px',
       accessor: 'id',
-      label: tWords('id'),
+      label: t('table.id'),
     },
     {
       maxW: '210px',
@@ -52,14 +47,14 @@ export const PatientCaseReportTable: FC<PatientCaseReportTableProps> = ({
       minW: '150px',
       textAlign: 'center',
       accessor: null,
-      label: tWords('actions'),
+      label: t('table.actions'),
       render: ({ data: caseReport }) => (
         <PatientCaseReportUpsertModal
           caseReportId={caseReport.id}
           patient={{ id: patientId, name: patientName }}
         >
           <Button size='sm' leftIcon={<EditIcon />}>
-            {tWords('update')}
+            {t('table.update')}
           </Button>
         </PatientCaseReportUpsertModal>
       ),
