@@ -1,16 +1,17 @@
+import { NextResponse } from 'next/server';
+
 import { EnumHttpStatus, sendError } from '~/core/api';
 import { getCustomerId } from '~/core/modules/Customer/auth';
 
 import { Patient } from '../model';
 import { ArchiveOrUnarchivePatientHandler } from './types';
 
-export const archive: ArchiveOrUnarchivePatientHandler = async (req, res) => {
-  const { authError, customerId } = await getCustomerId(req, res);
-  if (!customerId) return sendError({ res, error: authError });
+export const archive: ArchiveOrUnarchivePatientHandler = async (req, ctx) => {
+  const { authError, customerId } = await getCustomerId();
+  if (!customerId) return sendError({ error: authError });
 
-  const { patientId } = req.query;
-  if (!patientId)
-    return sendError({ res, error: 'patientId in query is required' });
+  const { patientId } = ctx.params;
+  if (!patientId) return sendError({ error: 'patientId in query is required' });
 
   await Patient.update(
     {
@@ -22,16 +23,15 @@ export const archive: ArchiveOrUnarchivePatientHandler = async (req, res) => {
     }
   );
 
-  res.status(EnumHttpStatus.Success).send({});
+  return NextResponse.json({}, { status: EnumHttpStatus.Success });
 };
 
-export const unarchive: ArchiveOrUnarchivePatientHandler = async (req, res) => {
-  const { authError, customerId } = await getCustomerId(req, res);
-  if (!customerId) return sendError({ res, error: authError });
+export const unarchive: ArchiveOrUnarchivePatientHandler = async (req, ctx) => {
+  const { authError, customerId } = await getCustomerId();
+  if (!customerId) return sendError({ error: authError });
 
-  const { patientId } = req.query;
-  if (!patientId)
-    return sendError({ res, error: 'patientId in query is required' });
+  const { patientId } = ctx.params;
+  if (!patientId) return sendError({ error: 'patientId in query is required' });
 
   await Patient.update(
     {
@@ -43,5 +43,5 @@ export const unarchive: ArchiveOrUnarchivePatientHandler = async (req, res) => {
     }
   );
 
-  res.status(EnumHttpStatus.Success).send({});
+  return NextResponse.json({}, { status: EnumHttpStatus.Success });
 };
