@@ -1,15 +1,17 @@
-import { EnumHttpStatus, sendError } from '~/core/api';
+import { NextResponse } from 'next/server';
+
+import { sendError } from '~/core/api';
 import { getCustomerId } from '~/core/modules/Customer/auth';
 
 import { Patient } from '../model';
 import { GetPatientHandler } from './types';
 
-export const get: GetPatientHandler = async (req, res) => {
-  const { authError, customerId } = await getCustomerId(req, res);
-  if (!customerId) return sendError({ res, error: authError });
+export const get: GetPatientHandler = async (req, ctx) => {
+  const { authError, customerId } = await getCustomerId();
+  if (!customerId) return sendError({ error: authError });
 
-  const { patientId: id } = req.query;
-  if (!id) return sendError({ res, error: 'patientId in query is required' });
+  const { patientId: id } = ctx.params;
+  if (!id) return sendError({ error: 'patientId in query is required' });
 
   const [patient] = await Patient.query({
     PK: { eq: customerId },
@@ -18,13 +20,12 @@ export const get: GetPatientHandler = async (req, res) => {
 
   if (!patient) {
     return sendError({
-      res,
       error: 'Patient not found',
       status: 'NotFound',
     });
   }
 
-  return res.status(EnumHttpStatus.Success).send({
+  return NextResponse.json({
     patient,
   });
 };
