@@ -1,10 +1,10 @@
 import { InferType } from 'yup';
 
 import { ReplaceNestedDates } from '../../../../../common/types/ReplaceNestedDates';
-import { patientSchema } from './schema';
+import { createPatientSchema } from './schema';
 
 export type PatientFieldsWithoutId = ReplaceNestedDates<
-  InferType<typeof patientSchema>
+  InferType<ReturnType<typeof createPatientSchema>>
 >;
 
 export type PatientFields = Omit<PatientFieldsWithoutId, 'id'> & {
