@@ -1,20 +1,18 @@
 import { FC } from 'react';
 
 import { Box, Text } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
 
 import { TextHighlight } from '~/common/components/TextHighlight';
 import { TooltipComponent } from '~/common/components/TooltipComponent';
 import { useQueryFilterByName } from '~/common/query';
 import { parseToHighlight } from '~/core/modules/Patient/api/parseSearchTerm';
+import { useScopedI18n } from '~/i18n/client';
 
 import { PatientNameProps } from './types';
 
 export const PatientName: FC<PatientNameProps> = ({ text, isArchived }) => {
   const { queryFilter } = useQueryFilterByName('patientName');
-  const { t } = useTranslation(undefined, {
-    keyPrefix: 'pages.patient.list.cell',
-  });
+  const t = useScopedI18n('translations.pages.patient.list.cell');
 
   return (
     <Box display='inline-block'>
