@@ -1,7 +1,6 @@
-import { useMemo } from 'react';
+import { useMemo, useRef } from 'react';
 
 import { Box, Flex, HStack } from '@chakra-ui/react';
-import { useTranslation } from 'react-i18next';
 
 import { LinkButton } from '~/common/components/Buttons';
 import {
@@ -10,12 +9,12 @@ import {
   FormSelectQueryFilter,
 } from '~/common/components/Form';
 import { Section } from '~/common/components/Section';
-import { schemaValidation } from '~/common/helpers';
 import { useDefaultQuery } from '~/common/query';
+import { schemaValidation } from '~/common/validation';
 import { Routes } from '~/core/constants';
-import { listPatientsSchema } from '~/core/modules/Patient/api/list/listPatientsSchema';
+import { createListPatientsSchema } from '~/core/modules/Patient/api/list/listPatientsSchema';
 import { PatientsTable } from '~/core/modules/Patient/pages/list/components/PatientsTable';
-import { AppPage } from '~/core/template/AppPage';
+import { useScopedI18n } from '~/i18n/client';
 
 import { EnumListPatientStatus } from '../../api/list/types';
 
@@ -30,9 +29,8 @@ const defaultQuery = {
 };
 
 export const ListPatients = () => {
-  const { t } = useTranslation(undefined, {
-    keyPrefix: 'pages.patient.list',
-  });
+  const schema = useRef(createListPatientsSchema());
+  const t = useScopedI18n('translations.pages.patient.list');
   const { getStateByString, stringParameters } = useDefaultQuery(defaultQuery);
 
   const patientStatusOptions = useMemo(
@@ -45,7 +43,7 @@ export const ListPatients = () => {
   );
 
   return (
-    <AppPage titleKey='patient.list'>
+    <Box w='100%'>
       <Section>
         <Flex alignItems='center'>
           <Box w='100%'>
@@ -53,7 +51,7 @@ export const ListPatients = () => {
               onSubmit={() => undefined}
               validateOnBlur
               initialValues={getStateByString(stringParameters)}
-              validate={schemaValidation(listPatientsSchema)}
+              validate={schemaValidation(schema.current)}
             >
               <HStack>
                 <Box w='100%' maxW={300}>
@@ -91,6 +89,6 @@ export const ListPatients = () => {
       </Section>
 
       <PatientsTable />
-    </AppPage>
+    </Box>
   );
 };
